Make numeric narrowing explicit.

+++ Option/BuildStationaryMap/buildstationarymap.cc
@@ -57,8 +57,8 @@ vector<shared_ptr<Option>> BuildStationaryMap::runPolicy(Writer<StringBuffer>& w
     updateStateObject();
 
   writer.String("ball").Bool(agentFrame->isBallVisible());
-  writer.String("goals").Uint(agentFrame->getGoalObservations().size());
-  writer.String("keepers").Uint(agentFrame->getTeamMateObservations().size());
+  writer.String("goals").Uint(static_cast<unsigned int>(agentFrame->getGoalObservations().size()));
+  writer.String("keepers").Uint(static_cast<unsigned int>(agentFrame->getTeamMateObservations().size()));
   writer.String("change").Bool(hasChange);
 
   return {};
